% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{apply_multiple_testing_methods}
\alias{apply_multiple_testing_methods}
\title{Run multiple testing procedures}
\usage{
apply_multiple_testing_methods(sim_df, alpha)
}
\arguments{
\item{sim_df}{`data.frame` returned from one of the simulation functions}

\item{alpha}{Nominal level at which to run the different multiple testing methods.}
}
\description{
Run multiple testing procedures
}
