% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bottomly}
\alias{bottomly}
\title{bottomly}
\format{
An object of class \code{RangedSummarizedExperiment} with 37991 rows and 21 columns.
}
\source{
<https://github.com/mikelove/bottomly>
}
\usage{
bottomly
}
\description{
Bottomly et al. RNASeq dataset.
}
\references{
Bottomly D, Walter NAR, Hunter JE, Darakjian P, Kawane S, Buck KJ, et al. (2011)
Evaluating Gene Expression in C57BL/6J and DBA/2J Mouse Striatum Using RNA-Seq and Microarrays.
PLoS ONE 6(3): e17820. https://doi.org/10.1371/journal.pone.0017820
}
\keyword{datasets}
