% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fisher_BH}
\alias{fisher_BH}
\title{Fisher combination followed by BH}
\usage{
fisher_BH(pvalues, pvalues2, alpha, Storey = FALSE)
}
\arguments{
\item{pvalues}{Numeric vector of unadjusted p-values.}

\item{pvalues2}{Numeric vector of secondary unadjusted p-values.}

\item{alpha}{Significance level at which to apply method}

\item{Storey}{Bool (default: FALSE): is the procedure pi0 adaptive or not?}
}
\value{
Binary vector of rejected/non-rejected hypotheses.
}
\description{
Fisher combination followed by BH
}
