% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{ihw_nmeth_wrapper}
\alias{ihw_nmeth_wrapper}
\title{Wrapper for the IHW procedure}
\usage{
ihw_nmeth_wrapper(Ps, Es, alpha, Storey = FALSE)
}
\arguments{
\item{Ps}{Numeric vector of unadjusted p-values.}

\item{Es}{Numeric vector of E-values}

\item{alpha}{Significance level at which to apply method}

\item{Storey}{Bool (default: FALSE): is the procedure pi0 adaptive or not?}
}
\value{
Binary vector of rejected/non-rejected hypotheses.
}
\description{
Wrapper for the IHW procedure
}
