% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalues.R
\name{limma_evalue}
\alias{limma_evalue}
\title{LIMMA (Replicated Microarray) E-value}
\usage{
limma_evalue(tstat, d, d0, gamma)
}
\arguments{
\item{tstat}{Moderated t-statistic.}

\item{d}{Likelihood degrees of freedom.}

\item{d0}{Prior degrees of freedom.}

\item{gamma}{Prior effect size variance scaling}
}
\value{
E-value
}
\description{
LIMMA (Replicated Microarray) E-value
}
