% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalues.R
\name{limma_results_table_with_evalues}
\alias{limma_results_table_with_evalues}
\title{Results table for LIMMA analysis that also includes e-values}
\usage{
limma_results_table_with_evalues(fit, coef)
}
\arguments{
\item{fit}{Object of class `MArrayLM` on which the `eBayes` method has already been applied to.
Typically this will be the output of the function `lmFit` followed by `eBayes`.}

\item{coef}{column name specifying which coefficient or contrast of the linear model is of interest}
}
\value{
A data.frame with the following columns:
   * t: moderated t-statistic
   * P.Value
   * B: the log odds ratio
   * evalue
}
\description{
Results table for LIMMA analysis that also includes e-values
}
