% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meanDispPairs}
\alias{meanDispPairs}
\title{meanDispPairs}
\format{
An object of class \code{data.frame} with 28501 rows and 2 columns.
}
\source{
<https://www.huber.embl.de/DESeq2paper/>
}
\usage{
meanDispPairs
}
\description{
Pairs of realistic mean and dispersion parameters used in the simulations
of the DESeq2 paper.
}
\references{
Love, M. I., Huber, W., & Anders, S. (2014).
Moderated estimation of fold change and dispersion for RNA-seq data with DESeq2.
Genome biology, 15(12), 1-21.
}
\keyword{datasets}
