% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{rnaseq_microarray_sim}
\alias{rnaseq_microarray_sim}
\title{RNA-Seq & Microarray simulation}
\usage{
rnaseq_microarray_sim(
  K,
  m,
  mean_disp_pairs,
  prop_signal_microarray = 0.5,
  es = 0.5,
  gamma = 1,
  prior_dof = 3.64,
  prior_s2 = 0.0144
)
}
\arguments{
\item{K}{Integer, number of hypotheses/genes}

\item{m}{Total number samples (which is assumed to be the same for RNA-Seq and microarray data).
Each group of the two-group comparison has m/2 samples.}

\item{mean_disp_pairs}{Dataframe with columns mean and disp that provide realistic values of the mean and dispersion of RNASeq count data.
See \link{meanDispPairs}.}

\item{prop_signal_microarray}{Number in [0, 1] (default: 0.5):What proportion
of genes differentially expressed in the RNA-Seq synthetic data are also
differentially expressed in the synthetic microarray data?}

\item{es}{Number, binary logarithmic fold change for differentially expressed genes in RNASeq data}

\item{gamma}{Number, determines strength of effect sizes for alternative genes (for synthetic microarray data)}

\item{prior_dof}{Number, Degrees of freedom for the prior of the inverse variances (for synthetic microarray data)}

\item{prior_s2}{Number, Prior variance}
}
\description{
A two-sample comparison based on both sources of information.
}
