% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{sim_mtp}
\alias{sim_mtp}
\title{Single index modulated multiple testing}
\usage{
sim_mtp(
  pvalues,
  pvalues2,
  alpha,
  grid_size = 40,
  details = FALSE,
  Storey = FALSE
)
}
\arguments{
\item{pvalues}{Numeric vector of unadjusted p-values.}

\item{pvalues2}{Numeric vector of secondary unadjusted p-values.}

\item{alpha}{Significance level at which to apply method}

\item{grid_size}{Integer (default: 40): How many gridpoints to choose to discretize the interval [0, pi/2] (corresponding to the possible angles between the two statistics)}

\item{details}{Bool (default: FALSE): Whether to return a detailed list of the output of the multiple testing procedure (e.g., including the optimal angle theta_opt)}

\item{Storey}{Bool (default: FALSE): is the procedure pi0 adaptive or not?}
}
\value{
Binary vector of rejected/non-rejected hypotheses.
}
\description{
Single index modulated multiple testing
}
