% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{single_sample_ttest_sim}
\alias{single_sample_ttest_sim}
\title{single sample t-test simulation}
\usage{
single_sample_ttest_sim(
  m,
  pi0,
  effect_size,
  n_samples = 10,
  var_halfrange = 0,
  evalue_ncp = 10,
  L = 6
)
}
\arguments{
\item{m}{Total number of hypotheses}

\item{pi0}{Proportion of nulls}

\item{effect_size}{Mean of alternative hypotheses}

\item{n_samples}{Total number samples used for each t-test. Each group has n_samples/2 samples.}

\item{evalue_ncp}{Noncentrality parameter used for calculation of e-values based on the likelihood ratio test}
}
\description{
Generates data of two sample t-test simulation
}
