% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tau_weighted_bh}
\alias{tau_weighted_bh}
\title{The tau-weighted BH multiple testing procedure}
\usage{
tau_weighted_bh(Ps, Ws, tau = 1, tau_pi0 = 0.5, Storey = FALSE)
}
\arguments{
\item{Ps}{Numeric vector of unadjusted p-values.}

\item{Ws}{Numeric vector of multiple testing weights}

\item{tau}{Numeric (default = 1), the level at which tau-censoring is applied. Forced to be <= tau_pi0 if Storey=TRUE.}

\item{tau_pi0}{Numeric (default = 0.5), the threshold at which Storey is applied}

\item{Storey}{Bool (default: FALSE): is the procedure pi0 adaptive or not?}
}
\value{
Vector of adjusted p-values
}
\description{
The tau-weighted BH multiple testing procedure
}
