% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{rnaseq_microarray_sim}
\alias{rnaseq_microarray_sim}
\title{RNA-Seq & Microarray simulation}
\usage{
rnaseq_microarray_sim(
  K,
  m,
  mean_disp_pairs,
  prop_signal_microarray = 0.5,
  es = 0.5,
  gamma = 1,
  prior_dof = 3.64,
  prior_s2 = 0.0144
)
}
\description{
RNA-Seq & Microarray simulation
}
