% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{single_sample_ttest_sim}
\alias{single_sample_ttest_sim}
\title{single sample t-test simulation}
\usage{
single_sample_ttest_sim(
  m,
  pi0,
  effect_size,
  n_samples = 10,
  var_halfrange = 0,
  evalue_ncp = 10,
  L = 6
)
}
\arguments{
\item{m}{Total number of hypotheses}

\item{pi0}{Proportion of nulls}

\item{effect_size}{Mean of alternative hypotheses}

\item{n_samples}{Total number samples used for each t-test.}

\item{var_halfrange}{The population variances of the data for each test are drawn from the uniform distribution on 1-var_halfrange, 1+var_halfrange. Defaults to var_halfrange=0, i.e., all population variances are equal to 1.}

\item{evalue_ncp}{Noncentrality parameter used for calculation of e-values based on the approximate likelihood ratio test.}

\item{L}{Highest monomial degree to keep in the Taylor expansion of the likelihood ratio e-value (default: L=6).}
}
\description{
Generates data of single sample t-test simulation
}
